requirements.txt
Django==2.1.5
django-alexa==0.0.9
python-dotenv==0.10.2
djangorestframework==3.9.0
django-ask-sdk==0.1.1